Extract the status bars out of App.ts
Every status surface in the panel - the status text, the Ready/Working/Error
pill, the progress bar, the error line, the diagnostics line and the home
screen dot - was written by a band of 347 lines sitting below renderApp. They
are pure DOM writes over AppElements with no generation state, no Photoshop and
no closure variables, so the whole group moves as a unit.

This is a relocation and nothing else: the moved lines are byte-identical to
the originals apart from the export keyword, verified by diffing the removed
band against the new file. The progress resolver test moves with it, from
src/ui/App to src/ui/statusBars.

App.ts drops from 5,842 to 5,513 lines.
